+++ src/main/java/com/jnape/palatable/lambda/functions/Effect.java
@@ -8,6 +8,7 @@ import java.util.function.Function;
 
 import static com.jnape.palatable.lambda.functions.Fn0.fn0;
 import static com.jnape.palatable.lambda.functions.IO.io;
+import static com.jnape.palatable.lambda.functions.builtin.fn1.Constantly.constantly;
 
 /**
  * A function returning "no result", and therefore only useful as a side-effect.
@@ -26,22 +27,22 @@ public interface Effect<A> extends Fn1<A, IO<Unit>>, Consumer<A> {
 
     @Override
     default <Z> Effect<Z> diMapL(Function<? super Z, ? extends A> fn) {
-        return effect(z -> Fn1.super.diMapL(fn).apply(z).unsafePerformIO());
+        return effect(Fn1.super.diMapL(fn));
     }
 
     @Override
     default <Z> Effect<Z> contraMap(Function<? super Z, ? extends A> fn) {
-        return effect(z -> Fn1.super.contraMap(fn).apply(z).unsafePerformIO());
+        return effect(Fn1.super.contraMap(fn));
     }
 
     @Override
     default <Z> Effect<Z> compose(Function<? super Z, ? extends A> before) {
-        return effect(z -> Fn1.super.compose(before).apply(z).unsafePerformIO());
+        return effect(Fn1.super.compose(before));
     }
 
     @Override
     default <C> Effect<A> discardR(Applicative<C, Fn1<A, ?>> appB) {
-        return effect(a -> Fn1.super.discardR(appB).apply(a).unsafePerformIO());
+        return effect(Fn1.super.discardR(appB));
     }
 
     @Override
@@ -67,6 +68,17 @@ public interface Effect<A> extends Fn1<A, IO<Unit>>, Consumer<A> {
      * @return the effect
      */
     static Effect<Unit> effect(Runnable runnable) {
-        return effect(__ -> runnable.run());
+        return effect(constantly(io(runnable)));
+    }
+
+    /**
+     * Create an {@link Effect} from an {@link Fn1} that yields an {@link IO}.
+     *
+     * @param fn  the function
+     * @param <A> the effect argument type
+     * @return the effect
+     */
+    static <A> Effect<A> effect(Fn1<A, ? extends IO<?>> fn) {
+        return a -> fn.apply(a).unsafePerformIO();
     }
 }