+++ src/test/java/com/jnape/palatable/lambda/functions/EffectTest.java
@@ -1,10 +1,14 @@
 package com.jnape.palatable.lambda.functions;
 
+import com.jnape.palatable.lambda.adt.Unit;
 import org.junit.Test;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.concurrent.atomic.AtomicInteger;
 
+import static com.jnape.palatable.lambda.adt.Unit.UNIT;
+import static com.jnape.palatable.lambda.functions.Effect.effect;
 import static com.jnape.palatable.lambda.functions.builtin.fn1.Constantly.constantly;
 import static java.util.Arrays.asList;
 import static org.junit.Assert.assertEquals;
@@ -30,5 +34,18 @@ public class EffectTest {
         andThen.accept("6");
 
         assertEquals(asList("1", "2", "3", "4", "5", "6", "6"), results);
+    }
+
+    @Test
+    public void staticFactoryMethods() {
+        AtomicInteger counter = new AtomicInteger();
+
+        Effect<Unit> runnableEffect = effect(counter::incrementAndGet);
+        runnableEffect.apply(UNIT).unsafePerformIO();
+        assertEquals(1, counter.get());
+
+        Effect<AtomicInteger> fnEffect = effect(AtomicInteger::incrementAndGet);
+        fnEffect.apply(counter).unsafePerformIO();
+        assertEquals(2, counter.get());
     }
 }